pyproject.toml
[tool.poetry]
name = "ChatBotRemember"
version = "1.0.0"
packages = [
    { include = "app" },
    { include = "database" },
]
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.setuptools.packages]
find = {}


[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.20.0.post0"
apscheduler = "3.11.0"
fastapi = "0.115.14"
pydantic = "2.11.7"
pydantic-settings = "2.10.1"
pyjwt = "2.10.1"
python-dotenv = "1.1.1"
sqlalchemy = "2.0.41"
uvicorn = "0.34.3"
alembic = "^1.16.2"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"